Conflicts:

	Documentation/feature-removal-schedule.txt
	arch/um/Kconfig.i386
	arch/um/Kconfig.x86_64
	arch/x86/kernel/acpi/Makefile
	arch/x86/kernel/acpi/sleep.h
	arch/x86/pci/acpi.c
	arch/x86/pci/common.c
	arch/x86/pci/pci.h
	include/asm-x86/ide.h
	include/asm-x86/semaphore_32.h
	include/asm-x86/semaphore_64.h
	include/asm-x86/trampoline.h
	lib/Kconfig.debug
	lib/Makefile

+++ include/linux/netdevice.h
@@ -193,8 +193,8 @@ struct dev_addr_list
 {
 	struct dev_addr_list	*next;
 	u8			da_addr[MAX_ADDR_LEN];
-	u8			da_addrlen;
-	u8			da_synced;
+	unsigned int		da_addrlen;
+	unsigned int		da_synced;
 	int			da_users;
 	int			da_gusers;
 };
